+++ backend/controllers/test.controller.js
@@ -19,6 +19,14 @@ const createTest = async (req, res) => {
   }
 
   try {
+    const exam = await Exam.findById({ _id: exam_id });
+    if (!exam) {
+      return res.status(400).json({
+        success: false,
+        msg: "Exam Not Found With This Exam Id",
+      });
+    }
+
     //create test entry in db
     const test = await Test.create({
       test_name,
@@ -29,7 +37,7 @@ const createTest = async (req, res) => {
     });
 
     // push test in exam
-    const exam = await Exam.findByIdAndUpdate(
+    const updatedExam = await Exam.findByIdAndUpdate(
       { _id: exam_id },
       { $push: { all_tests: test._id } },
       { new: true }
@@ -40,7 +48,7 @@ const createTest = async (req, res) => {
       success: "true",
       msg: "Test Created Successfully",
       test,
-      exam,
+      updatedExam,
     });
   } catch (error) {
     return res.status(500).json({
@@ -57,7 +65,7 @@ const editTest = async (req, res) => {
   const { test_name, total_questions, total_marks, total_time, test_id } =
     req.body;
 
-    console.log(test_id)
+  console.log(test_id);
 
   // check test is avalable or not
   const test = await Test.findById({ _id: test_id });
@@ -108,7 +116,7 @@ const showAllTests = async (req, res) => {
     return res.status(200).json({
       success: "true",
       msg: "Got All Tests Successfully",
-      allTests
+      allTests,
     });
   } catch (error) {
     return res.status(500).json({
@@ -119,52 +127,54 @@ const showAllTests = async (req, res) => {
   }
 };
 
-
 //=================================================Delete Test============================================
 const deleteTest = async (req, res) => {
-    const {test_id, exam_id} = req.body
+  const { test_id, exam_id } = req.body;
 
-    try {
-      //Find test
-      const test = await Test.findById({_id:test_id})
-      if(!test){
-        return res.status(401).json({
-          success: "false",
-          msg: "Test not found with this Id",
-        });
-      }
-
-      // find all questions answer and options and delete them
-      const allQuestions = test.questions;
-        for (const questionId of allQuestions) {
-          const question = await Question.findById({ _id: questionId });
-          if (question) {
-            const allOptions = question.options;
-            for (const optionId of allOptions) {
-              // delete All Options
-              await Option.findByIdAndDelete({ _id: optionId });
-            }
-            const answerId = question.answer;
-            // delete all answer
-            await Answer.findByIdAndDelete({ _id: answerId });
-          }
-
-          // delete all questions
-          await Question.findByIdAndDelete({ _id: questionId });
-        }
-
-        // remove test id from exam 
-        await Exam.findByIdAndUpdate({_id: exam_id},
-          {$pull : {all_tests : test_id}}
-          )
-
-
-    } catch (error) {
-      return res.status(500).json({
+  try {
+    //Find test
+    const test = await Test.findById({ _id: test_id });
+    if (!test) {
+      return res.status(401).json({
         success: "false",
-        msg: "Something Went Wrong",
-        error: error.message
+        msg: "Test not found with this Id",
       });
     }
-}
+
+    // find all questions answer and options and delete them
+    const allQuestions = test.questions;
+    for (const questionId of allQuestions) {
+      const question = await Question.findById({ _id: questionId });
+      if (question) {
+        const allOptions = question.options;
+        for (const optionId of allOptions) {
+          // delete All Options
+          if(optionId){
+            await Option.findByIdAndDelete({ _id: optionId });
+          }
+        }
+        const answerId = question.answer;
+        // delete all answer
+        if(answerId){
+          await Answer.findByIdAndDelete({ _id: answerId });
+        }
+      }
+
+      // delete all questions
+      await Question.findByIdAndDelete({ _id: questionId });
+    }
+
+    // remove test id from exam
+    await Exam.findByIdAndUpdate(
+      { _id: exam_id },
+      { $pull: { all_tests: test_id } }
+    );
+  } catch (error) {
+    return res.status(500).json({
+      success: "false",
+      msg: "Something Went Wrong",
+      error: error.message,
+    });
+  }
+};
 export { createTest, editTest, showAllTests, deleteTest };
